add: `MockCoreWrapper` 入力反映出力
