+++ voicevox_engine/dev/core/mock.py
@@ -67,8 +67,11 @@ class MockCoreWrapper(CoreWrapper):
         self, length: int, phoneme_list: ndarray, style_id: ndarray
     ) -> ndarray:
         """音素系列サイズ・音素ID系列・スタイルIDから音素長系列を生成する"""
-        # Mock: 定数の音素長系列を生成。[0.1, 0.1, ...]
-        return 0.1 * numpy.ones((length,), dtype=numpy.float32)
+        result = []
+        # mockとしての適当な処理、特に意味はない
+        for i in range(length):
+            result.append(round((phoneme_list[i] * 0.0625 + style_id).item(), 2))
+        return numpy.array(result)
 
     def yukarin_sa_forward(
         self,
@@ -83,12 +86,28 @@ class MockCoreWrapper(CoreWrapper):
     ) -> ndarray:
         """モーラ系列サイズ・母音系列・子音系列・アクセント位置・アクセント句区切り・スタイルIDからモーラ音高系列を生成する"""
         assert length > 1, "前後無音を必ず付与しなければならない"
-        # Mock: 定数のモーラ音高系列を生成。[0, 200, 100, 100, ..., 100, 0]
-        pitch = 100 * numpy.ones((1, length), dtype=numpy.float32)
-        pitch[0, 0] = 0.0  # 開始無音 (pau)
-        pitch[0, 1] = 200.0  # 分散 0 を避けるため
-        pitch[0, length - 1] = 0.0  # 終了無音 (pau)
-        return pitch
+
+        result = []
+        # mockとしての適当な処理、特に意味はない
+        for i in range(length):
+            result.append(
+                round(
+                    (
+                        (
+                            vowel_phoneme_list[0][i]
+                            + consonant_phoneme_list[0][i]
+                            + start_accent_list[0][i]
+                            + end_accent_list[0][i]
+                            + start_accent_phrase_list[0][i]
+                            + end_accent_phrase_list[0][i]
+                        )
+                        * 0.0625
+                        + style_id
+                    ).item(),
+                    2,
+                )
+            )
+        return numpy.array(result)[numpy.newaxis]
 
     def decode_forward(
         self,
@@ -99,8 +118,18 @@ class MockCoreWrapper(CoreWrapper):
         style_id: ndarray,
     ) -> ndarray:
         """フレーム長・音素種類数・フレーム音高・フレーム音素onehot・スタイルIDから音声波形を生成する"""
-        # Mock: 定数の音声波形を生成。[0.1, 0.1, ..., 0.1, 0.1]
-        return 0.1 * numpy.ones((length * 256,), dtype=numpy.float32)
+        result = []
+        # mockとしての適当な処理、特に意味はない
+        for i in range(length):
+            # decode forwardはデータサイズがlengthの256倍になるのでとりあえず256回データをresultに入れる
+            for _ in range(256):
+                result.append(
+                    (
+                        f0[i][0] * (numpy.where(phoneme[i] == 1)[0] / phoneme_size)
+                        + style_id
+                    ).item()
+                )
+        return numpy.array(result)
 
     def supported_devices(self):
         return json.dumps(
